+++ packages/account-tree-controller/src/logger.ts
@@ -1,0 +1,9 @@
+import { createProjectLogger, createModuleLogger } from '@metamask/utils';
+
+import { controllerName } from './AccountTreeController';
+
+export const projectLogger = createProjectLogger(controllerName);
+export const backupAndSyncLogger = createModuleLogger(
+  projectLogger,
+  'Backup and sync',
+);