+++ packages/ens-controller/src/EnsController.ts
@@ -82,8 +82,18 @@ export type EnsControllerMessenger = RestrictedMessenger<
 >;
 
 const metadata = {
-  ensEntries: { persist: true, anonymous: false },
-  ensResolutionsByAddress: { persist: true, anonymous: false },
+  ensEntries: {
+    includeInStateLogs: true,
+    persist: true,
+    anonymous: false,
+    usedInUi: true,
+  },
+  ensResolutionsByAddress: {
+    includeInStateLogs: true,
+    persist: true,
+    anonymous: false,
+    usedInUi: true,
+  },
 };
 
 const defaultState = {